refactor: Move DummyFormulaActor to utils, remove useless actors from tests.utils, clean code
