+++ tests/acceptation/crash/test_crash_dispatcher.py
@@ -58,12 +58,12 @@ import signal
 import os
 import pytest
 
-from powerapi.formula import DummyFormulaActor
 
 from tests.utils.acceptation import MainProcess
 from tests.utils.acceptation import get_actor_by_name, DISPATCHER_ACTOR_NAME
 # noinspection PyUnresolvedReferences
 from tests.utils.db.mongo import mongo_database
+from tests.utils.formula.dummy import DummyFormulaActor
 from tests.utils.report.hwpc import extract_rapl_reports_with_2_sockets
