+++ tests/unit/puller/simple/test_simple_puller_actor.py
@@ -38,7 +38,7 @@ from powerapi.message import StartMessage, \
 from powerapi.filter import Filter
 from powerapi.report import HWPCReport
 from powerapi.puller.simple.simple_puller_actor import SimplePullerActor
-from tests.utils.dummy_actor import DummyActor
+from tests.utils.actor.dummy_actor import DummyActor
 
 from tests.unit.actor.abstract_test_actor import AbstractTestActor, recv_from_pipe, is_actor_alive
