+++ tests/utils/actor/dummy_actor.py
@@ -1,21 +1,21 @@
 # Copyright (c) 2022, INRIA
 # Copyright (c) 2022, University of Lille
 # All rights reserved.
-#
+
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
-#
+
 # * Redistributions of source code must retain the above copyright notice, this
 #   list of conditions and the following disclaimer.
-#
+
 # * Redistributions in binary form must reproduce the above copyright notice,
 #   this list of conditions and the following disclaimer in the documentation
 #   and/or other materials provided with the distribution.
-#
+
 # * Neither the name of the copyright holder nor the names of its
 #   contributors may be used to endorse or promote products derived from
 #   this software without specific prior written permission.
-#
+
 # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 # AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 # IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -27,4 +27,39 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from powerapi.formula.dummy.dummy_formula_actor import DummyFormulaActor, DummyFormulaState
+from powerapi.handler import PoisonPillMessageHandler, StartHandler
+from powerapi.message import StartMessage, PoisonPillMessage
+from powerapi.actor import Actor, State
+from tests.utils.actor.dummy_handlers import DummyHandler
+
+LOGGER_NAME = 'multiprocess_test_logger'
+
+
+class DummyActorState(State):
+    """
+    Message used to start a DummyActor
+    :param pipe: pipe used to send message to pytest process
+    """
+
+    def __init__(self, actor: Actor, pipe):
+        State.__init__(self, actor)
+        self.pipe = pipe
+
+
+class DummyActor(Actor):
+    """
+    Actor that forward all start message (except StartMessage) into a pipe to the test process
+    """
+
+    def __init__(self, name: str, pipe, message_type):
+        Actor.__init__(self, name)
+        self.state = DummyActorState(self, pipe)
+        self.message_type = message_type
+
+    def setup(self):
+        """
+        Define message handler
+        """
+        self.add_handler(self.message_type, DummyHandler(self.state))
+        self.add_handler(PoisonPillMessage, PoisonPillMessageHandler(self.state))
+        self.add_handler(StartMessage, StartHandler(self.state))
