+++ tests/utils/formula/__init__.py
@@ -1,5 +1,5 @@
-# Copyright (c) 2022, INRIA
-# Copyright (c) 2022, University of Lille
+# Copyright (c) 2023, INRIA
+# Copyright (c) 2023, University of Lille
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -26,8 +26,3 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-from powerapi.formula.handlers import FormulaPoisonPillMessageHandler
-from powerapi.formula.abstract_cpu_dram_formula import AbstractCpuDramFormula
-from powerapi.formula.formula_actor import FormulaActor, FormulaState
-from powerapi.formula.simple.simple_formula_actor import SimpleFormulaActor
